Add `credentials` provider and auth models
