+++ src/features/auth/models/auth.d.ts
@@ -1,18 +1,5 @@
-import type { User } from 'next-auth';
-type UserId = string;
-
-declare module 'next-auth/jwt' {
-  // Extenfing JWT interface to include user id
-
-  interface JWT {
-    id: UserId;
-  }
-}
-
-declare module 'next-auth' {
-  // Overriding User type to include id
-
-  interface Session {
-    user: User & { id: UserId };
-  }
+interface Token {
+  id: UserId;
+  email: string;
+  exp: number;
 }
