+++ src/features/auth/providers/credentials.ts
@@ -1,0 +1,22 @@
+import CredentialsProvider from 'next-auth/providers/credentials';
+
+export default CredentialsProvider({
+  id: 'credentials',
+  name: 'credentials',
+  credentials: {
+    email: { label: 'Email', type: 'email' },
+    password: { label: 'Password', type: 'password' },
+    token: { label: 'Token', type: 'text' },
+  },
+  async authorize(credentials) {
+    // if (!credentials) {
+    //   throw AuthError.invalidCredentials();
+    // }
+    // if (credentials.token) {
+    //   return await loginWithToken(credentials.token);
+    // }
+    // return await login(credentials.email, credentials.password);
+
+    return null;
+  },
+});
